Moved Khezu config
